+++ src/features/helper/types/variables.ts
@@ -1,15 +1,15 @@
 type Variables = {
-  workspaceFolder: string;
-  workspaceFolderBasename: string;
-  file: string;
-  relativeFile: string;
-  fileBasename: string;
-  fileBasenameNoExtension: string;
-  fileDirname: string;
-  fileExtname: string;
-  lineNumber: number;
-  selectedText: string;
-  execPath: string;
+  workspaceFolder?: string;
+  workspaceFolderBasename?: string;
+  file?: string;
+  relativeFile?: string;
+  fileBasename?: string;
+  fileBasenameNoExtension?: string;
+  fileDirname?: string;
+  fileExtname?: string;
+  lineNumber?: string;
+  selectedText?: string;
+  execPath?: string;
 };
 
 export default Variables;